Doing some Pointer/Pointed stuff

+++ cinterop/jni/JNIImpls.cpp
@@ -1,5 +1,9 @@
+// ReSharper disable CppCStyleCast
+// ReSharper disable CppRedundantCastExpression
+#include <cstdint>
+
 #include "io_github_epicvon2468_cinterop_jni_JNIImpls.h"
 
 JNIEXPORT jlong JNICALL Java_io_github_epicvon2468_cinterop_jni_JNIImpls_getNativeNullPointer(JNIEnv* env, jclass clazz) {
-	return NULL;
+	return (long)(intptr_t)NULL;
 }
